backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic[email]
passlib[bcrypt]
python-jose
httpx
pytest
